# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MikroTik Centralized Fleet Management API
  version: 1.20.0
...
servers:
- url: /api/v1
paths:
...
  /dashboard/global:
    get:
      summary: Global dashboard metrics
      tags:
      - Dashboard
      parameters:
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalDashboardResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /dashboard/global/reconciliation:
    get:
      summary: Tong data + do lech (gap) toan ham doi
      description: Gop cung cong thuc computeGap() dang dung o cap tung tau (ships/{shipId}/reconciliation) len toan ham doi -- tong byte WAN + CREW/BUSINESS/MANAGEMENT cong don qua tat ca tau, va breakdown do lech tung tau sap theo |gap| giam dan. 422 neu hoan toan khong co interface_counter_deltas nao trong ca ham doi cho khoang thoi gian nay.
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalReconciliationResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /areas/{areaId}/dashboard:
    get:
      summary: Area dashboard metrics
      tags:
      - Dashboard
      parameters:
      - $ref: '#/components/parameters/AreaId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: OK. Inventory is available even when telemetry is insufficient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaDashboardResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: AREA_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /ships/{shipId}/dashboard:
    get:
      summary: Ship dashboard
      tags:
      - Dashboard
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipDashboardResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation:
    get:
      summary: Ship data reconciliation summary
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: Reconciliation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation/by-wan:
    get:
      summary: Reconciliation grouped by WAN input
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      responses:
        '200':
          description: Reconciliation by WAN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationByWanResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation/by-interface:
    get:
      summary: Reconciliation grouped by interface
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      responses:
        '200':
          description: Reconciliation by interface
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationByInterfaceResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation/by-zone:
    get:
      summary: Reconciliation grouped by network zone
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: Reconciliation by zone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationByZoneResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation/raw-records:
    get:
      summary: Raw reconciliation records drill-down
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      - $ref: '#/components/parameters/RawLimit'
      - $ref: '#/components/parameters/RawCursor'
      - $ref: '#/components/parameters/RawSource'
      - $ref: '#/components/parameters/RawReason'
      responses:
        '200':
          description: Raw records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationRawRecordsResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/reconciliation/timeseries:
    get:
      summary: Reconciliation gap time series
      tags:
      - Reconciliation
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      - $ref: '#/components/parameters/DashboardZone'
      responses:
        '200':
          description: Reconciliation time series
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationTimeseriesResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RECONCILIATION_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /ships/{shipId}/crew/users:
    get:
      summary: CREW user list with RADIUS-derived usage for a period
      tags:
      - Crew
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - &id020
        $ref: '#/components/parameters/DashboardFrom'
      - &id021
        $ref: '#/components/parameters/DashboardTo'
      - &id022
        $ref: '#/components/parameters/DashboardTimezone'
      - &id023
        $ref: '#/components/parameters/DashboardGranularity'
      - in: header
        name: x-actor-permissions
        required: false
        description: Temporary AuthStubGuard permission header; production authentication is not implemented in this phase.
        schema:
          type: string
      responses:
        '200':
          description: CREW users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewUsersResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/crew/users/{username}/sessions:
    get:
      summary: RADIUS session drill-down for one CREW user
      tags:
      - Crew
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/Username'
      - $ref: '#/components/parameters/RawCursor'
      - $ref: '#/components/parameters/RawLimit'
      - in: header
        name: x-actor-permissions
        required: false
        description: Temporary AuthStubGuard permission header; production authentication is not implemented in this phase.
        schema:
          type: string
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewSessionsResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ships/{shipId}/crew/users/{username}/sessions/{sessionId}/disconnect:
    post:
      summary: Ngat 1 phien RADIUS dang ACTIVE (CoA/Disconnect-Request that, RFC 5176)
      description: Gui Disconnect-Request toi NAS (router) dang giu phien nay. KHONG tu danh dau session CLOSED o day -- trang thai that chi doi khi Accounting-Stop that su toi. Router phai da bat "RADIUS incoming" (accept-from chua IP backend), neu khong se TIMEOUT.
      tags:
      - Crew
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - $ref: '#/components/parameters/Username'
      - in: path
        name: sessionId
        required: true
        schema:
          type: string
          format: uuid
      - in: header
        name: x-actor-permissions
        required: false
        schema:
          type: string
      responses:
        '201':
          description: NAS da xac nhan (Disconnect-ACK)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      acknowledged:
                        type: boolean
                      message:
                        type: string
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    nullable: true
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND hoac RADIUS_SESSION_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: RADIUS_SESSION_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: RADIUS_DISCONNECT_FAILED -- NAS tu choi (NAK), timeout, hoac loi mang
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /ships/{shipId}/crew/raw-accounting:
    get:
      summary: Raw RADIUS accounting event drill-down
      tags:
      - Crew
      parameters:
      - $ref: '#/components/parameters/ShipId'
      - *id020
      - *id021
      - *id022
      - *id023
      - in: query
        name: username
        required: false
        schema:
          type: string
      - in: query
        name: acct_session_id
        required: false
        schema:
          type: string
      - $ref: '#/components/parameters/RawCursor'
      - $ref: '#/components/parameters/RawLimit'
      - in: header
        name: x-actor-permissions
        required: false
        description: Temporary AuthStubGuard permission header; production authentication is not implemented in this phase.
        schema:
          type: string
      responses:
        '200':
          description: Raw accounting records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewRawAccountingResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AreaId:
      in: path
      name: areaId
      required: true
      schema:
        type: string
        format: uuid
    ShipId:
      in: path
      name: shipId
      required: true
      schema:
        type: string
        format: uuid
    DashboardFrom:
      in: query
      name: from
      required: false
      schema:
        type: string
        format: date-time
      description: RFC3339 period start. Defaults to 24 hours before `to`.
    DashboardTo:
      in: query
      name: to
      required: false
      schema:
        type: string
        format: date-time
      description: RFC3339 period end. Defaults to now.
    DashboardTimezone:
      in: query
      name: timezone
      required: false
      schema:
        type: string
      description: IANA timezone used for period bucketing.
    DashboardGranularity:
      in: query
      name: granularity
      required: false
      schema:
        type: string
        enum:
        - 1m
        - 5m
        - 1h
        - 1d
        default: 1h
    DashboardZone:
      in: query
      name: zone
      required: false
      schema:
        type: string
        enum:
        - CREW
        - BUSINESS
        - MANAGEMENT
        - ALL
        default: ALL
...
    RawLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    RawCursor:
      in: query
      name: cursor
      required: false
      schema:
        type: string
    RawSource:
      in: query
      name: source
      required: false
      schema:
        type: string
        enum:
        - INTERFACE_COUNTER
        - RADIUS_ACCOUNTING
        - IPFIX
        - DNS
        - SYSLOG
    RawReason:
      in: query
      name: reason
      required: false
      schema:
        type: string
    Username:
      in: path
      name: username
      required: true
      schema:
        type: string
  schemas:
...
    GlobalReconciliationResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            period:
              $ref: '#/components/schemas/DashboardPeriod'
            units:
              $ref: '#/components/schemas/DashboardUnits'
            data_status:
              type: string
              enum:
              - AVAILABLE
            formula_version:
              type: string
            ship_count:
              type: integer
              description: So tau THAT SU co it nhat 1 dong interface_counter_deltas trong khoang thoi gian nay (khong phai tong so tau trong inventory).
            fleet:
              type: object
              properties:
                total_bytes:
                  type: integer
                  nullable: true
                  description: "\"Tong data\" toan ham doi = WAN download + upload cong lai cho ca ky."
                wan:
                  $ref: '#/components/schemas/GlobalReconciliationZoneBytes'
                crew:
                  type: object
                  properties:
                    port_download_bytes:
                      type: integer
                      nullable: true
                    port_upload_bytes:
                      type: integer
                      nullable: true
                    user_download_bytes:
                      type: integer
                      nullable: true
                    user_upload_bytes:
                      type: integer
                      nullable: true
                business:
                  $ref: '#/components/schemas/GlobalReconciliationZoneBytes'
                management:
                  $ref: '#/components/schemas/GlobalReconciliationZoneBytes'
                gaps:
                  $ref: '#/components/schemas/GlobalReconciliationGaps'
            ships:
              type: array
              items:
                $ref: '#/components/schemas/GlobalReconciliationShipRow'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    GlobalDashboardResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            scope:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - GLOBAL
            period:
              $ref: '#/components/schemas/DashboardPeriod'
            units:
              $ref: '#/components/schemas/DashboardUnits'
            data_status:
              type: string
              enum:
              - AVAILABLE
              - INSUFFICIENT_DATA
            availability:
              $ref: '#/components/schemas/DashboardAvailability'
            data_quality:
              $ref: '#/components/schemas/DataQuality'
            total_areas:
              type: integer
            total_ships:
              type: integer
            total_devices:
              type: integer
            online_ships:
              type: integer
            total_wan_throughput:
              type: integer
              nullable: true
            active_crew_users:
              type: integer
              nullable: true
            degraded_ships:
              type: integer
            offline_ships:
              type: integer
            traffic:
              $ref: '#/components/schemas/DashboardTrafficSummary'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    ShipDashboardResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            scope:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - SHIP
                id:
                  type: string
            ship:
              $ref: '#/components/schemas/ShipDashboardIdentity'
            period:
              $ref: '#/components/schemas/DashboardPeriod'
            units:
              $ref: '#/components/schemas/DashboardUnits'
            data_status:
              type: string
              enum:
              - AVAILABLE
              - INSUFFICIENT_DATA
            availability:
              $ref: '#/components/schemas/DashboardAvailability'
            data_quality:
              $ref: '#/components/schemas/DataQuality'
            connectivity:
              type: object
              additionalProperties: true
            wan:
              type: array
              items:
                type: object
                additionalProperties: true
            crew:
              type: object
              additionalProperties: true
            business:
              type: object
              additionalProperties: true
            reconciliation_summary:
              type: object
              additionalProperties: true
            alerts:
              type: object
              additionalProperties: true
            status:
              type: string
            wan_rx:
              type: integer
              nullable: true
            wan_tx:
              type: integer
              nullable: true
            active_crew_users:
              type: integer
              nullable: true
            crew_usage:
              type: integer
              nullable: true
            business_usage:
              type: integer
              nullable: true
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
    AreaDashboardResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            scope:
              type: object
              additionalProperties: true
            area:
              type: object
              additionalProperties: true
            period:
              $ref: '#/components/schemas/DashboardPeriod'
            units:
              $ref: '#/components/schemas/DashboardUnits'
            data_status:
              type: string
              enum:
              - AVAILABLE
              - INSUFFICIENT_DATA
            availability:
              $ref: '#/components/schemas/DashboardAvailability'
            data_quality:
              $ref: '#/components/schemas/DataQuality'
            inventory:
              type: object
              additionalProperties: true
            ships:
              type: array
              items:
                $ref: '#/components/schemas/AreaDashboardShip'
            traffic:
              $ref: '#/components/schemas/DashboardTrafficSummary'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    DashboardTrafficSummary:
      type: object
      properties:
        wan_download_bytes:
          type: integer
          nullable: true
        wan_upload_bytes:
          type: integer
          nullable: true
        crew_bytes:
          type: integer
          nullable: true
        business_bytes:
          type: integer
          nullable: true
...
    ReconciliationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationData'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    ReconciliationByWanResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationItemsData'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
    ReconciliationByInterfaceResponse:
      $ref: '#/components/schemas/ReconciliationByWanResponse'
    ReconciliationByZoneResponse:
      $ref: '#/components/schemas/ReconciliationByWanResponse'
    ReconciliationTimeseriesResponse:
      $ref: '#/components/schemas/ReconciliationByWanResponse'
    ReconciliationRawRecordsResponse:
      type: object
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/ReconciliationItemsData'
          - type: object
            properties:
              page:
                type: object
                additionalProperties: true
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          $ref: '#/components/schemas/ErrorDetails'
...
    CrewUsersResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            period:
              $ref: '#/components/schemas/DashboardPeriod'
            units:
              type: object
              additionalProperties: true
            source:
              type: string
            data_status:
              type: string
              enum:
              - AVAILABLE
              - INSUFFICIENT_DATA
            data_quality:
              $ref: '#/components/schemas/ModuleDataQuality'
            identity_capability:
              $ref: '#/components/schemas/IdentityCapability'
            service_usage_capability:
              type: object
              additionalProperties: true
            stale_after_seconds:
              type: integer
            billing_reconciliation:
              type: object
              nullable: true
              description: Compares NetFlow-measured bytes (attributed to a known CREW identity) against RADIUS-reported bytes for the same ship/period — the "billing" anchor per the reference design's coverage check. Null when there is no RADIUS traffic to compare against.
              properties:
                netflow_total_bytes:
                  type: integer
                  nullable: true
                billing_total_bytes:
                  type: integer
                  nullable: true
                coverage_pct:
                  type: number
                  nullable: true
            users:
              type: array
              items:
                $ref: '#/components/schemas/CrewUser'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...
    CrewSessionsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            username:
              type: string
            units:
              type: object
              additionalProperties: true
            source:
              type: string
            identity_capability:
              $ref: '#/components/schemas/IdentityCapability'
            sessions:
              type: array
              items:
                $ref: '#/components/schemas/CrewSession'
        meta:
          $ref: '#/components/schemas/BaseMeta'
        error:
          nullable: true
...